update location service, dto

+++ DTOS/LocationDto.cs
@@ -5,8 +5,8 @@ namespace barangay_crime_compliant_api.DTOS
     {
 
         public long Id { get; set; }
-        public decimal? Lat { get; set; }
-        public decimal? Long { get; set; }
+        public float? Lat { get; set; }
+        public float? Long { get; set; }
         public string? Description { get; set; }
         public DateTime? DateTimeCreated { get; set; }
         public DateTime? DateTimeUpdated { get; set; }
